+++ CHANGELOG.rst
@@ -2,11 +2,18 @@
 Changelog
 =========
 
+0.6.2 (2022-02-24)
+==================
+
 * Added percentage type for thresholds
 * Added ``datasource`` parameter to CloudWatch targets
 * Added support for auto panels ids to AlertList panel
+* Added ``SeriesOverride`` options (dashes and Z-index)
 * Added support for fields value in Stat panel
 * Added ``alertName`` parameter to AlertList panel
+* Added ``thresholdsStyleMode`` parameter to TimeSeries panel
+* Added Histogram panel support
+* Dashboard upload script updated to support overwriting dashboards
 
 0.6.1 (2021-11-23)
 ==================
@@ -52,7 +59,7 @@ Changes
 
 * Added colour overrides to pie chart panel
 * Added missing attributes from xAxis class
-* Added transformations for the Panel class (https://grafana.com/docs/grafana/next/panels/transformations/types-options/#transformation-types-and-options)
+* Added transformations for the Panel class (https://grafana.com/docs/grafana/next/panels/reference-transformation-functions/)
 * Added Worldmap panel (https://grafana.com/grafana/plugins/grafana-worldmap-panel/)
 * Added missing fill gradient to Graph panel
 * Added missing align to graph panel
@@ -124,7 +131,7 @@ Changes
 
 * Added Alert Threshold enabled/disabled to Graphs.
 * Added constants for all Grafana value formats
-* Added support for repetitions to Stat Panels (https://grafana.com/docs/grafana/latest/variables/repeat-panels-or-rows/)
+* Added support for repetitions to Stat Panels
 * Added textMode option to Stat Panels
 * Add Panel object for all panels to inherit from
 * Add Dashboard list panel (https://grafana.com/docs/grafana/latest/panels/visualizations/dashboard-list-panel/)